requirements.txt
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
comtypes==1.4.12
fastapi==0.117.1
greenlet==3.2.4
idna==3.10
mysql-connector-python==9.4.0
psycopg2-binary==2.9.10
PyAudio==0.2.14
pydantic==2.11.9
pydantic_core==2.33.2
pypiwin32==223
pyttsx==1.1
pyttsx3==2.99
pywin32==311
setuptools==80.9.0
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
